+++ tests/unit/components/purchase/ApproveSpendingStep.spec.ts
@@ -1,15 +1,24 @@
 import { mount, createLocalVue, Wrapper } from '@vue/test-utils'
 import VueRouter from 'vue-router'
+import { router } from '../../../../src/router'
+import flushPromises from 'flush-promises'
 
 import { getModule } from 'vuex-module-decorators'
 import appStore from '../../../../src/store'
 import PurchaseModule from '../../../../src/vuexModules/PurchaseModule'
 import AppModule from '../../../../src/vuexModules/AppModule'
+import EventModule from '../../../../src/vuexModules/EventModule'
 
 import ApproveSpendingStep from '@/components/purchase/ApproveSpendingStep.vue'
 
 import { ProcessStatus } from '../../../../src/models/ProcessStatus'
 import { PurchaseStep } from '../../../../src/models/PurchaseStep'
+
+
+import PurchaseProcessModule from '../../../../src/functionModules/components/PurchaseProcessModule'
+import EventableModule from '../../../../src/functionModules/eventable/EventableModule'
+import EtherTokenContractModule from '../../../../src/functionModules/protocol/EtherTokenContractModule'
+
 
 import { library } from '@fortawesome/fontawesome-svg-core'
 import { faSpinner } from '@fortawesome/free-solid-svg-icons'
@@ -27,8 +36,9 @@ describe('ApproveSpendingStep.vue', () => {
 
   const localVue = createLocalVue()
 
-  let appModule!: AppModule
   let purchaseModule!: PurchaseModule
+  let appModule!: AppModule
+  let eventModule!: EventModule
   let wrapper!: Wrapper<ApproveSpendingStep>
 
   beforeAll(() => {
@@ -40,6 +50,8 @@ describe('ApproveSpendingStep.vue', () => {
     appModule = getModule(AppModule, appStore)
     purchaseModule = getModule(PurchaseModule, appStore)
 
+    eventModule = getModule(EventModule, appStore)
+
     appModule.setDatatrustContractAllowance(allowance)
     purchaseModule.setStatus(ProcessStatus.Ready)
   })
@@ -48,6 +60,45 @@ describe('ApproveSpendingStep.vue', () => {
     if (wrapper !== undefined) {
       wrapper.destroy()
     }
+  })
+
+  it('Approves the datatrust, if needed', async () => {
+    purchaseModule.setPurchaseStep(PurchaseStep.ApproveSpending)
+
+    EtherTokenContractModule.allowance = () => Promise.resolve('10')
+    EtherTokenContractModule.approve = () => Promise.resolve()
+    EtherTokenContractModule.balanceOf = () => Promise.resolve('1000000000')
+    PurchaseProcessModule.getPurchasePrice = () => 100
+
+    wrapper = mount(ApproveSpendingStep, {
+      attachToDocument: true,
+      store: appStore,
+      localVue,
+      router,
+    })
+
+    const wrapTokenButtonDiv = wrapper.find('.approve-spending .process-button')
+    const wrapTokenButton = wrapper.find('.process-button .button')
+
+
+    // Initiate wrapping transaction
+    wrapTokenButton.trigger('click')
+    expect(purchaseModule.purchaseStep).toBe(PurchaseStep.ApprovalPending)
+    expect(wrapTokenButtonDiv.vm.$props.clickable).toBeTruthy()
+
+    const minedProcessId = purchaseModule.approvalMinedProcessId
+
+    // Update new allowance amount
+    EtherTokenContractModule.allowance = () => Promise.resolve('1000')
+
+    // create an event signifying mining finsihed
+    eventModule.append(EventableModule.createEvent(minedProcessId, true , undefined))
+
+    await flushPromises()
+
+    // purchase step is now to approve, button is no longer clickable
+    expect(purchaseModule.purchaseStep).toBe(PurchaseStep.PurchaseListing)
+    expect(wrapTokenButtonDiv.vm.$props.clickable).toBeFalsy()
   })
 
   // NOTE: Removed conditional rendering for now, instead made conditionally clickable
